django2fa/requirements.txt
django-two-factor-auth==1.17.0
phonenumbers==8.13.55
pillow==11.1.0
platformdirs==4.2.2
pyotp==2.9.0
tomli==2.0.1
webauthn==2.5.2
